aider: Added buttons for each drink choice in the `index.html` file.
# Aider chat conversation:

USER: This SPA is for tracking alcohol consumption. I want buttons that the user will click to add a drink. The drink choices are beer glass, beer pint, big beer, Glass of wine, cocktail

ASSISTANT: To add buttons for each drink choice, we can modify the body of your `index.html` file. Each button will have a unique id and a click event that will trigger a function to add a drink.

Here is the edit block for the `index.html` file:

```html
index.html
<<<<<<< HEAD
<body>
    <!-- Your htmx code goes here -->
</body>
=======
<body>
    <!-- Your htmx code goes here -->
    <button id="beerGlass" onclick="addDrink('beerGlass')">Beer Glass</button>
    <button id="beerPint" onclick="addDrink('beerPint')">Beer Pint</button>
    <button id="bigBeer" onclick="addDrink('bigBeer')">Big Beer</button>
    <button id="wineGlass" onclick="addDrink('wineGlass')">Glass of Wine</button>
    <button id="cocktail" onclick="addDrink('cocktail')">Cocktail</button>
</body>
>>>>>>> updated
```

This will add five buttons to your HTML file. When a button is clicked, it will call the `addDrink` function with the id of the button as a parameter.

Please note that the `addDrink` function is not defined yet. You will need to define this function in your JavaScript file to handle the drink addition. If you want me to help with that, please add your JavaScript file to the chat.