aider: Define `addDrink` function to log the drink type to the console.
# Aider chat conversation:

USER: Write addDrink

ASSISTANT: Sure, we can define the `addDrink` function in your `app.js` file. This function will take the drink type as a parameter and add the drink to a list of drinks. For now, let's just log the drink type to the console.

Here is the edit block for the `app.js` file:

```javascript
app.js
<<<<<<< HEAD
// Your JavaScript code will go here
=======
// Your JavaScript code will go here
function addDrink(drinkType) {
    console.log(drinkType);
}
>>>>>>> updated
```

This will define the `addDrink` function. When a button is clicked, the id of the button will be logged to the console.

Please note that this is a very basic implementation of the `addDrink` function. Depending on your application's requirements, you might want to update this function to add the drink to a list, update the UI, etc.