aider: Added script tag to link `app.js` file to `index.html`.
# Aider chat conversation:

USER: Is my `app.js` connected to `index.html`?

ASSISTANT: Based on the provided `index.html` file, it seems that the `app.js` file is not yet linked. To link the `app.js` file, you need to add a script tag at the end of the body in the `index.html` file.

Here is the edit block for the `index.html` file:

```html
index.html
<<<<<<< HEAD
</body>
</html>
=======
    <script src="app.js"></script>
</body>
</html>
>>>>>>> updated
```

This will link the `app.js` file to the `index.html` file. Now, the `addDrink` function defined in the `app.js` file will be accessible in the `index.html` file.