aider: Added spinner input for alcohol percent and modified addDrink function to calculate alcohol intake based on drink type and alcohol percent.
# Aider chat conversation:

USER: Okay, so each of those drinks has an alcohol amount associated with it. The tracking formula is simple, it is simply alocohol ounces x alcohol percent. So a 12oz glass of 5% abv beer would be 12*5=60. Lets modify so that when you push a button, the daily alcohol intake tally is incremented by the amount of that drink. To allow the user to specify the alcohol percent, we need a spinner with a range of 1-60.

ASSISTANT: To implement this, we need to make a few changes:

1. Add a spinner input to the `index.html` file to allow the user to specify the alcohol percent.
2. Modify the `addDrink` function in the `app.js` file to calculate the alcohol intake based on the drink type and the alcohol percent. We will also need to add a variable to keep track of the daily alcohol intake.

Here are the edit blocks for these changes:

```html
index.html
<<<<<<< HEAD
    <!-- Your htmx code goes here -->
    <button id="beerGlass" onclick="addDrink('beerGlass')">Beer Glass</button>
    <button id="beerPint" onclick="addDrink('beerPint')">Beer Pint</button>
    <button id="bigBeer" onclick="addDrink('bigBeer')">Big Beer</button>
    <button id="wineGlass" onclick="addDrink('wineGlass')">Glass of Wine</button>
    <button id="cocktail" onclick="addDrink('cocktail')">Cocktail</button>
=======
    <!-- Your htmx code goes here -->
    <input type="number" id="alcoholPercent" min="1" max="60" value="5">
    <button id="beerGlass" onclick="addDrink('beerGlass', document.getElementById('alcoholPercent').value)">Beer Glass</button>
    <button id="beerPint" onclick="addDrink('beerPint', document.getElementById('alcoholPercent').value)">Beer Pint</button>
    <button id="bigBeer" onclick="addDrink('bigBeer', document.getElementById('alcoholPercent').value)">Big Beer</button>
    <button id="wineGlass" onclick="addDrink('wineGlass', document.getElementById('alcoholPercent').value)">Glass of Wine</button>
    <button id="cocktail" onclick="addDrink('cocktail', document.getElementById('alcoholPercent').value)">Cocktail</button>
>>>>>>> updated
```

```javascript
app.js
<<<<<<< HEAD
// Your JavaScript code will go here
function addDrink(drinkType) {
    console.log(drinkType);
}
=======
// Your JavaScript code will go here
var dailyAlcoholIntake = 0;

function addDrink(drinkType, alcoholPercent) {
    var alcoholOunces;
    switch(drinkType) {
        case 'beerGlass':
            alcoholOunces = 12;
            break;
        case 'beerPint':
            alcoholOunces = 16;
            break;
        case 'bigBeer':
            alcoholOunces = 22;
            break;
        case 'wineGlass':
            alcoholOunces = 5;
            break;
        case 'cocktail':
            alcoholOunces = 1.5;
            break;
    }
    dailyAlcoholIntake += alcoholOunces * alcoholPercent;
    console.log('Daily alcohol intake: ' + dailyAlcoholIntake);
}
>>>>>>> updated
```

In the `index.html` file, we added a spinner input for the alcohol percent and modified the `onclick` attribute of each button to pass the alcohol percent to the `addDrink` function.

In the `app.js` file, we added a `dailyAlcoholIntake` variable to keep track of the daily alcohol intake. We also modified the `addDrink` function to calculate the alcohol intake based on the drink type and the alcohol percent. The alcohol intake is then added to the `dailyAlcoholIntake` variable and logged to the console.