aider: Changed alcoholOunces value for the 'cocktail' case in the addDrink function from 1.5 to 3.
# Aider chat conversation:

USER: Let's assume that the base `alcoholOunces` in a cocktail is 3oz.

ASSISTANT: To reflect this change, we need to modify the `addDrink` function in the `app.js` file. Specifically, we need to change the `alcoholOunces` value for the 'cocktail' case in the switch statement from 1.5 to 3.

Here is the *edit block* for this change:

```javascript
app.js
<<<<<<< HEAD
    case 'cocktail':
        alcoholOunces = 1.5;
        break;
=======
    case 'cocktail':
        alcoholOunces = 3;
        break;
>>>>>>> updated
```