aider: Added a new div element to display the current alcohol count and modified the addDrink function to update the content of the div element.
# Aider chat conversation:

USER: Can we prominantly display the current alcohol count?

ASSISTANT: Sure, we can display the current alcohol count prominently on the page. We will need to modify both the `index.html` and `app.js` files.

In the `index.html` file, we will add a new `div` element to display the current alcohol count.

In the `app.js` file, we will modify the `addDrink` function to update the content of the new `div` element every time a drink is added.

Here are the changes:

```html
index.html
<<<<<<< HEAD
    <button id="cocktail" onclick="addDrink('cocktail', document.getElementById('alcoholPercent').value)">Cocktail</button>
    <script src="app.js"></script>
</body>
</html>
=======
    <button id="cocktail" onclick="addDrink('cocktail', document.getElementById('alcoholPercent').value)">Cocktail</button>
    <div id="alcoholCount">Current Alcohol Count: 0</div>
    <script src="app.js"></script>
</body>
</html>
>>>>>>> updated
```

```javascript
app.js
<<<<<<< HEAD
    dailyAlcoholIntake += alcoholOunces * alcoholPercent;
    console.log('Daily alcohol intake: ' + dailyAlcoholIntake);
=======
    dailyAlcoholIntake += alcoholOunces * alcoholPercent;
    document.getElementById('alcoholCount').textContent = 'Current Alcohol Count: ' + dailyAlcoholIntake;
>>>>>>> updated
```