aider: Change the cocktail `alcoholOunces` to 2.
# Aider chat conversation:

USER: Let's change the cocktail `alcoholOunces` to 2

ASSISTANT: To change the `alcoholOunces` for the 'cocktail' case in the `addDrink` function, we need to modify the `app.js` file. Here is the required change:

```javascript
app.js
<<<<<<< HEAD
    case 'cocktail':
        alcoholOunces = 3;
        break;
=======
    case 'cocktail':
        alcoholOunces = 2;
        break;
>>>>>>> updated
```

This change will reduce the alcohol content for the 'cocktail' case from 3 ounces to 2 ounces.