aider: Added link to `style.css` in `index.html`.
# Aider chat conversation:

USER: Have we associated `style.css` with `index.html`?

ASSISTANT: Based on the provided `index.html` file, it appears that the `style.css` file is not currently linked. To apply the styles from `style.css` to your HTML document, you need to link the CSS file in the `head` section of your `index.html` file.

Here is the change that needs to be made:

```html
index.html
<<<<<<< HEAD
    <title>AlcoChart</title>
    <link rel="manifest" href="/manifest.json">
    <script src="https://unpkg.com/htmx.org@1.6.1"></script>
=======
    <title>AlcoChart</title>
    <link rel="stylesheet" type="text/css" href="style.css">
    <link rel="manifest" href="/manifest.json">
    <script src="https://unpkg.com/htmx.org@1.6.1"></script>
>>>>>>> updated
```

This change adds a link to the `style.css` file in the `head` section of your `index.html` file. The `rel="stylesheet"` attribute specifies that the linked file is a stylesheet, and the `type="text/css"` attribute specifies that the file is a CSS file. The `href="style.css"` attribute specifies the path to the CSS file.