aider: Fixed issue with `setPreset` function not finding elements with IDs `drinkSize` and `alcoholStrength` by wrapping the function in a `window.onload` event handler.
# Aider chat conversation:

USER:     at setPreset (app.js:30:48)

ASSISTANT: The error message indicates that the issue is in the `setPreset` function in the `app.js` file, specifically at the line where we're trying to set the `value` property of an element that doesn't exist.

The `setPreset` function is trying to update the `value` property of the elements with the IDs `drinkSize` and `alcoholStrength`. However, these elements are not being found, hence the error.

This could be due to the fact that the `setPreset` function is being called before the DOM is fully loaded, and as a result, the elements with the IDs `drinkSize` and `alcoholStrength` are not available at the time the function is called.

To fix this, we can wrap the code that calls the `setPreset` function in a `window.onload` event handler. This will ensure that the `setPreset` function is not called until the entire page, including all dependent resources such as stylesheets and images, have finished loading.

Here's the *edit block* for the changes:

```javascript
app.js
<<<<<<< HEAD
function setPreset(preset) {
    switch(preset) {
        case 'beerGlass':
            drinkSize = 12;
            alcoholStrength = 5;
            break;
        case 'beerPint':
            drinkSize = 16;
            alcoholStrength = 5;
            break;
        case 'bigBeer':
            drinkSize = 20;
            alcoholStrength = 5;
            break;
        case 'wineGlass':
            drinkSize = 5;
            alcoholStrength = 12;
            break;
        case 'cocktail':
            drinkSize = 2;
            alcoholStrength = 40;
            break;
    }
    document.getElementById('drinkSize').value = drinkSize;
    document.getElementById('alcoholStrength').value = alcoholStrength;
    updateAlcoholCount();
}
=======
window.onload = function() {
    function setPreset(preset) {
        switch(preset) {
            case 'beerGlass':
                drinkSize = 12;
                alcoholStrength = 5;
                break;
            case 'beerPint':
                drinkSize = 16;
                alcoholStrength = 5;
                break;
            case 'bigBeer':
                drinkSize = 20;
                alcoholStrength = 5;
                break;
            case 'wineGlass':
                drinkSize = 5;
                alcoholStrength = 12;
                break;
            case 'cocktail':
                drinkSize = 2;
                alcoholStrength = 40;
                break;
        }
        document.getElementById('drinkSize').value = drinkSize;
        document.getElementById('alcoholStrength').value = alcoholStrength;
        updateAlcoholCount();
    }
}
>>>>>>> updated
```

This *edit block* wraps the `setPreset` function in a `window.onload` event handler, ensuring that it's not called until the entire page has finished loading.