aider: Rotate vertical axis labels 90 degrees to make them run vertically.
# Aider chat conversation:

USER: Better. The vertical axis grid labels are not running vertically though. They are running horizontally, which is the wrong way.

ASSISTANT: To make the vertical axis labels run vertically, we need to rotate them 90 degrees. This can be done by modifying the `#yAxisLabels` rule in the `style.css` file.

Here is the *edit block* for this change:

```css
style.css
<<<<<<< HEAD
#yAxisLabels {
    display: flex;
    justify-content: space-between;
    width: 100%;
    position: absolute;
    left: 0;
    transform: rotate(0deg);
}
=======
#yAxisLabels {
    display: flex;
    justify-content: space-between;
    width: 100%;
    position: absolute;
    left: 0;
    transform: rotate(-90deg);
}
>>>>>>> updated
```

Please apply this change to your `style.css` file.