aider: Reversed the order of the labels in the y-axis.
# Aider chat conversation:

USER: Now they are almost correct. They are increasing from top-down though, rather than bottom-up.

ASSISTANT: I see, the labels are being displayed in the reverse order. This is because the `flex-direction: column;` property in the `#yAxisLabels` rule displays the labels from top to bottom. We can reverse this by adding `flex-direction: column-reverse;` to the `#yAxisLabels` rule.

Here is the *edit block* for this change:

```css
style.css
<<<<<<< HEAD
#yAxisLabels {
    display: flex;
    flex-direction: column;
    justify-content: space-between;
    height: 100%;
    position: absolute;
    left: 0;
    overflow: visible;
}
=======
#yAxisLabels {
    display: flex;
    flex-direction: column-reverse;
    justify-content: space-between;
    height: 100%;
    position: absolute;
    left: 0;
    overflow: visible;
}
>>>>>>> updated
```

Please apply this change to your `style.css` file.